aiassistant/backend/requirements.txt
Flask>=3.0.3
flask-cors>=4.0.1
pymupdf>=1.24.11
pdfplumber>=0.11.4
faiss-cpu>=1.9.0
sentence-transformers>=3.3.0
pymongo[srv]>=4.10.1
PyJWT>=2.10.0
werkzeug>=3.1.2
groq>=0.13.0
numpy>=2.1.0
paddleocr>=2.7.3
paddlepaddle>=2.6.1
groq>=0.13.0
python-pptx>=1.0.2
python-docx>=1.1.0
python-dotenv>=1.0.1
gunicorn>=21.2.0
